refactor: 리뷰 사항 반영
RacingCarRace.race -> RacingCar.startRace
