+++ src/main/kotlin/racingcar/RacingCarExecuteMain.kt
@@ -1,5 +1,5 @@
 package racingcar
 
 fun main() {
-    RacingCarRace().race()
+    RacingCar.startRace()
 }
