+++ src/test/kotlin/racingcar/RacingCarTest.kt
@@ -9,23 +9,23 @@ class RacingCarTest {
     @Test
     fun `무작위 값 테스트`() {
         for (i in 1..100) {
-            val randInt = RacingCarMove.createRandomInt()
+            val randInt = RacingCar.createRandomInt()
             assertTrue(randInt in 0..9)
         }
     }
 
     @Test
     fun `전진 조건 테스트`() {
-        assertTrue(RacingCarMove.checkMoveValue(5))
+        assertTrue(RacingCar.checkMoveValue(5))
     }
 
     @Test
     fun `전진 조건이 아닌 경우 테스트`() {
-        assertFalse(RacingCarMove.checkMoveValue(3))
+        assertFalse(RacingCar.checkMoveValue(3))
     }
 
     @Test
     fun `전진 문자 테스트`() {
-        assertTrue(RacingCarMove.getMoveString(3) == "---")
+        assertTrue(RacingCar.getMoveString(3) == "---")
     }
 }
